Fix overflow at 2^32 nodes
git-svn-id: svn+ssh://keith.cs.st-and.ac.uk/mnt/raid/svn/repositories/minion-private/public/trunk@674 477b24dd-e91e-0410-aa49-980f3df0729b

+++ minion/solver2.h
@@ -29,7 +29,7 @@
 
 VARDEF_ASSIGN(clock_t time_limit, 0);
 
-VARDEF(unsigned nodes);
+VARDEF(unsigned long long nodes);
 
 
 namespace Controller
@@ -51,7 +51,7 @@ namespace Controller
   VARDEF_ASSIGN(int commandlineoption_sollimit, -1);
   VARDEF_ASSIGN(BOOL commandlineoption_fullpropogate, false);
   VARDEF_ASSIGN(BOOL commandlineoption_nocheck, false);
-  VARDEF_ASSIGN(unsigned commandlineoption_nodelimit, 0);
+  VARDEF_ASSIGN(unsigned long long commandlineoption_nodelimit, 0);
   VARDEF(vector<int> test_solution);
